Update Bogus remove Randomizer.Seed usage
Conflicts:
	src/Tests/Framework/MockData/Generators.cs
