+++ src/Tests/Framework/MockData/Gimme.cs
@@ -1,12 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Bogus;
 
 namespace Tests.Framework.MockData
 {
 	internal static class Gimme
 	{
-		public static Randomizer Random = new Randomizer(TestClient.Configuration.Seed);
+		public static Randomizer Random { get; } = new Randomizer(TestClient.Configuration.Seed);
 	}
 }
