+++ src/Tests/Framework/MockData/Ranges.cs
@@ -106,6 +106,7 @@ namespace Tests.Framework.MockData
 
 		public static Faker<Ranges> Generator { get; } =
 			new Faker<Ranges>()
+				.UseSeed(TestClient.Configuration.Seed)
 				.CustomInstantiator((f) => new Ranges(f))
 			;
 	}
