+++ src/Tests/Framework/MockData/SimpleGeoPoint.cs
@@ -9,6 +9,7 @@ namespace Tests.Framework.MockData
 
 		public static Faker<SimpleGeoPoint> Generator { get; } =
 			new Faker<SimpleGeoPoint>()
+				.UseSeed(TestClient.Configuration.Seed)
 				.RuleFor(p => p.Lat, f => f.Address.Latitude())
 				.RuleFor(p => p.Lon, f => f.Address.Longitude())
 			.Clone();
