+++ src/Tests/Framework/TestClient.cs
@@ -2,7 +2,6 @@
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Text;
 using Elastic.Managed.Configuration;
@@ -46,7 +45,7 @@ namespace Tests.Framework
 	public static class TestClient
 	{
 		public static readonly bool RunningFiddler = Process.GetProcessesByName("fiddler").Any();
-		public static readonly ITestConfiguration Configuration = LoadConfiguration();
+		public static readonly ITestConfiguration Configuration = ConfigurationLoader.LoadConfiguration();
 
 		public static readonly ConnectionSettings GlobalDefaultSettings = CreateSettings();
 		public static readonly IElasticClient Default = new ElasticClient(GlobalDefaultSettings);
@@ -59,34 +58,6 @@ namespace Tests.Framework
 
 		public static string DefaultHost => "localhost";
 		private static string Host => (RunningFiddler) ? "ipv4.fiddler" : DefaultHost;
-
-		private static ITestConfiguration LoadConfiguration()
-		{
-			// The build script sets a FAKEBUILD env variable, so if it exists then
-			// we must be running tests from the build script
-			if (!string.IsNullOrEmpty(Environment.GetEnvironmentVariable("FAKEBUILD")))
-				return new EnvironmentConfiguration();
-
-			var directoryInfo = new DirectoryInfo(Directory.GetCurrentDirectory());
-
-			// If running the classic .NET solution, tests run from bin/{config} directory,
-			// but when running DNX solution, tests run from the test project root
-			var yamlConfigurationPath = (directoryInfo.Name == "Tests"
-				&& directoryInfo.Parent != null
-				&& directoryInfo.Parent.Name == "src")
-				? "."
-				: @"../../../";
-
-			var localYamlFile = Path.GetFullPath(Path.Combine(yamlConfigurationPath, "tests.yaml"));
-			if (File.Exists(localYamlFile))
-				return new YamlConfiguration(localYamlFile);
-
-			var defaultYamlFile = Path.GetFullPath(Path.Combine(yamlConfigurationPath, "tests.default.yaml"));
-			if (File.Exists(defaultYamlFile))
-				return new YamlConfiguration(defaultYamlFile);
-
-			throw new Exception($"Tried to load a yaml file from {yamlConfigurationPath} but it does not exist : pwd:{directoryInfo.FullName}");
-		}
 
 		private static int ConnectionLimitDefault =>
 			int.TryParse(Environment.GetEnvironmentVariable("NEST_NUMBER_OF_CONNECTIONS"), out int x)
